+++ lib/extension/src/object_extension.dart
@@ -360,5 +360,5 @@ extension ExtensionT<T> on T {
     return this;
   }
 
-  List<T> toList() => [this];
+  List<T> convertToList() => [this];
 }