+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.3.0+5
+
+* Add `convertToList()` for Type T
+
 ## 3.3.0+3
 
 * Add `DecoratorBoxState()` and Example